feat: group similar position and arrange properties
