+++ frontend/src/components/AlignmentControls.vue
@@ -1,29 +1,6 @@
 <template>
-	<CollapsibleSection title="Alignment">
+	<CollapsibleSection title="Alignment" :initialState="false">
 		<template #default>
-			<Button label="Backward" @click="sendBackward" />
-			<Button label="Back" @click="sendToBack" />
-			<Button label="Forward" @click="sendForward" />
-			<Button label="Front" @click="sendToFront" />
-			<div class="flex items-center gap-3">
-				<NumberInput
-					:modelValue="selectionBounds.left"
-					@update:modelValue="(val) => updatePosition('X', val)"
-					prefix="x"
-					:rangeStart="0"
-					:rangeStep="1"
-					:hideButtons="true"
-				/>
-				<NumberInput
-					:modelValue="selectionBounds.top"
-					@update:modelValue="(val) => updatePosition('Y', val)"
-					prefix="y"
-					:rangeStart="0"
-					:rangeStep="1"
-					:hideButtons="true"
-				/>
-			</div>
-
 			<div v-for="axis in axes" :key="axis" class="flex flex-col gap-1.5">
 				<div :class="fieldLabelClasses">{{ axis.label }}</div>
 				<div class="grid grid-cols-3 gap-3">
@@ -57,16 +34,9 @@ import {
 
 import CollapsibleSection from '@/components/controls/CollapsibleSection.vue'
 
-import { slideBounds, selectionBounds, guideVisibilityMap, currentSlide } from '@/stores/slide'
+import { slideBounds, selectionBounds, guideVisibilityMap } from '@/stores/slide'
 import { fieldLabelClasses } from '@/utils/constants'
-import {
-	activeElements,
-	getElementPosition,
-	isWithinOverlappingBounds,
-	normalizeZIndices,
-} from '@/stores/element'
-
-import { cloneObj } from '@/utils/helpers'
+import { updatePosition } from '@/stores/element'
 
 const horizontalAlignmentOptions = [
 	{
@@ -161,131 +131,5 @@ const performAlignment = (direction) => {
 
 const updateGuideVisibilityMap = (direction, value) => {
 	guideVisibilityMap[direction] = value
-}
-
-const updatePosition = (axis, value) => {
-	const property = axis == 'X' ? 'left' : 'top'
-
-	const delta = value - selectionBounds[property]
-
-	activeElements.value.forEach((element) => {
-		element[property] += delta
-	})
-
-	selectionBounds[property] = value
-}
-
-const moveElement = (elements, elementId, moveToIndex, action) => {
-	const movingElement = elements.find((el) => el.id == elementId)
-
-	elements.forEach((el) => {
-		if (
-			action.includes('back') &&
-			el.zIndex >= moveToIndex &&
-			el.zIndex < movingElement.zIndex
-		) {
-			el.zIndex += 1
-		} else {
-			if (el.zIndex <= moveToIndex && el.zIndex > movingElement.zIndex) {
-				el.zIndex -= 1
-			}
-		}
-	})
-
-	movingElement.zIndex = moveToIndex
-}
-
-const getElementLists = (action) => {
-	// use cloned objects so changes are applied all at once
-	// for cleaner history updation
-	const elements = cloneObj(currentSlide.value.elements)
-	const active = cloneObj(activeElements.value)
-
-	const sortedActiveElements = ['back', 'backward'].includes(action)
-		? active.sort((a, b) => a.zIndex - b.zIndex)
-		: active.sort((a, b) => b.zIndex - a.zIndex)
-
-	return {
-		elements,
-		sortedActiveElements,
-	}
-}
-
-const isElementWithinBounds = (activeId, elementId) => {
-	const activePosition = getElementPosition(activeId)
-	const elementPosition = getElementPosition(elementId)
-
-	return isWithinOverlappingBounds(activePosition, elementPosition)
-}
-
-const initMoveToIndexAndFactor = (elements, sortedActiveElements, action) => {
-	const baseIndex = sortedActiveElements[0].zIndex
-	let moveToIndex = null
-
-	const isOverlappingElement = (el) => {
-		const isBackward = action == 'backward' && el.zIndex < baseIndex
-		const isForward = action == 'forward' && el.zIndex > baseIndex
-
-		if (isBackward || isForward) {
-			return isElementWithinBounds(sortedActiveElements[0].id, el.id)
-		}
-		return false
-	}
-
-	switch (action) {
-		case 'back':
-			return { moveToIndex: 1, factor: 1 }
-		case 'front':
-			return { moveToIndex: elements.length, factor: -1 }
-		case 'backward':
-			const lowerZIndices = elements
-				.filter((el) => isOverlappingElement(el))
-				.map((el) => el.zIndex)
-			return {
-				moveToIndex: lowerZIndices.length ? Math.max(...lowerZIndices) : 1,
-				factor: 1,
-			}
-		case 'forward':
-			const higherZIndices = elements
-				.filter((el) => isOverlappingElement(el))
-				.map((el) => el.zIndex)
-			return {
-				moveToIndex: higherZIndices.length ? Math.min(...higherZIndices) : elements.length,
-				factor: -1,
-			}
-		default:
-			return { moveToIndex: baseIndex, factor: 1 }
-	}
-}
-
-const getElementsWithUpdatedZIndices = (action) => {
-	const { elements, sortedActiveElements } = getElementLists(action)
-
-	let { moveToIndex, factor } = initMoveToIndexAndFactor(elements, sortedActiveElements, action)
-
-	sortedActiveElements.forEach((element) => {
-		moveElement(elements, element.id, moveToIndex, action)
-
-		// next element will move one position above the previous one
-		moveToIndex += factor
-	})
-
-	return normalizeZIndices(elements)
-}
-
-const sendBackward = () => {
-	currentSlide.value.elements = getElementsWithUpdatedZIndices('backward')
-}
-
-const sendToBack = () => {
-	currentSlide.value.elements = getElementsWithUpdatedZIndices('back')
-}
-
-const sendForward = () => {
-	currentSlide.value.elements = getElementsWithUpdatedZIndices('forward')
-}
-
-const sendToFront = () => {
-	currentSlide.value.elements = getElementsWithUpdatedZIndices('front')
 }
 </script>
