+++ frontend/src/components/MediaProperties.vue
@@ -85,18 +85,6 @@
 			/>
 		</template>
 	</CollapsibleSection>
-
-	<CollapsibleSection title="Other" :initialState="false">
-		<template #default>
-			<SliderInput
-				label="Opacity"
-				:rangeStart="0"
-				:rangeEnd="100"
-				:modelValue="activeElement.opacity"
-				@update:modelValue="(value) => setProperty('opacity', value)"
-			/>
-		</template>
-	</CollapsibleSection>
 </template>
 
 <script setup>
@@ -104,6 +92,7 @@ import SliderInput from '@/components/controls/SliderInput.vue'
 import NumberInput from '@/components/controls/NumberInput.vue'
 import ColorPicker from '@/components/controls/ColorPicker.vue'
 import CollapsibleSection from '@/components/controls/CollapsibleSection.vue'
+import PositionArrangementProperties from './PositionArrangementProperties.vue'
 
 import { activeElement } from '@/stores/element'
 import { fieldLabelClasses } from '@/utils/constants'
