+++ frontend/src/icons/Backward.vue
@@ -1,0 +1,26 @@
+<template>
+	<svg
+		xmlns="http://www.w3.org/2000/svg"
+		width="16"
+		height="16"
+		viewBox="0 0 24 24"
+		fill="none"
+		stroke="currentColor"
+		stroke-width="1.5"
+		stroke-linecap="round"
+		stroke-linejoin="round"
+		class="lucide lucide-layers2-icon lucide-layers-2"
+	>
+		<path
+			d="M13 13.74a2 2 0 0 1-2 0L2.5 8.87a1 1 0 0 1 0-1.74L11 2.26a2 2 0 0 1 2 0l8.5 4.87a1 1 0 0 1 0 1.74z"
+		/>
+		<path
+			d="M13 19.74a2 2 0 0 1-2 0L2.5 14.87a1 1 0 0 1 0-1.74L11 8.26a2 2 0 0 1 2 0l8.5 4.87a1 1 0 0 1 0 1.74z"
+			fill="#e6e6e6"
+		/>
+	</svg>
+</template>
+
+<script setup></script>
+
+<style lang="scss" scoped></style>
