+++ frontend/src/icons/BringToFront.vue
@@ -1,0 +1,22 @@
+<template>
+	<svg
+		xmlns="http://www.w3.org/2000/svg"
+		width="16"
+		height="16"
+		viewBox="0 0 24 24"
+		fill="none"
+		stroke="currentColor"
+		stroke-width="1.5"
+		stroke-linecap="round"
+		stroke-linejoin="round"
+		class="lucide lucide-bring-to-front-icon lucide-bring-to-front"
+	>
+		<rect x="6.5" y="6.5" width="11" height="11" rx="2" fill="#e6e6e6" />
+		<path d="M4 10a2 2 0 0 1-2-2V4a2 2 0 0 1 2-2h4a2 2 0 0 1 2 2" />
+		<path d="M14 20a2 2 0 0 0 2 2h4a2 2 0 0 0 2-2v-4a2 2 0 0 0-2-2" />
+	</svg>
+</template>
+
+<script setup></script>
+
+<style lang="scss" scoped></style>
