+++ frontend/src/icons/SendToBack.vue
@@ -1,0 +1,22 @@
+<template>
+	<svg
+		xmlns="http://www.w3.org/2000/svg"
+		width="16"
+		height="16"
+		viewBox="0 0 24 24"
+		fill="none"
+		stroke="currentColor"
+		stroke-width="1.5"
+		stroke-linecap="round"
+		stroke-linejoin="round"
+		class="lucide lucide-send-to-back-icon lucide-send-to-back"
+	>
+		<rect x="6.5" y="6.5" width="11" height="11" rx="2" fill="#e6e6e6" />
+		<rect x="14" y="14" width="8" height="8" rx="2" fill="white" />
+		<rect x="2" y="2" width="8" height="8" rx="2" fill="white" />
+	</svg>
+</template>
+
+<script setup></script>
+
+<style lang="scss" scoped></style>
